Create helper to factorize code
